Support block size that are multiples of 32 in DP4A Matmul
